# Optimize App Startup by Background Loading of Firebase Remote Config