--- a/lib/remote_configuration/lib/src/implementation/firebase_remote_configuration.dart
+++ b/lib/remote_configuration/lib/src/implementation/firebase_remote_configuration.dart
@@ -6,6 +6,7 @@
 //
 // SPDX-License-Identifier: EUPL-1.2
 
+import 'dart:async';
 import 'dart:developer';
 
 import 'package:firebase_remote_config/firebase_remote_config.dart';
@@ -29,15 +30,33 @@ class FirebaseRemoteConfiguration extends RemoteConfiguration {
     return _remoteConfig.getBool(key);
   }
 
+  /// Initializes the remote configuration with the given default values and
+  /// fetches the remote configuration in the background.
+  ///
+  /// The default values are used if the remote configuration could not be
+  /// fetched.
+  ///
+  /// This method follows the "Load new values for next startup" strategy:
+  /// https://firebase.google.com/docs/remote-config/loading.
   @override
-  Future<void> initialize(Map<String, dynamic> defaultValues) async {
+  Future<void> initializeAndFetchInBackground(
+    Map<String, dynamic> defaultValues,
+  ) async {
     try {
       _remoteConfig = FirebaseRemoteConfig.instance;
       _remoteConfig.setDefaults(defaultValues);
-      _remoteConfig.setConfigSettings(RemoteConfigSettings(
+      _remoteConfig.setConfigSettings(
+        RemoteConfigSettings(
           fetchTimeout: const Duration(minutes: 1),
-          minimumFetchInterval: const Duration(hours: 3)));
-      await _remoteConfig.fetchAndActivate();
+          minimumFetchInterval: const Duration(hours: 3),
+        ),
+      );
+      // Activate the fetched remote config from the last fetch.
+      await _remoteConfig.activate();
+
+      // Fetch remote config in the background. They will be available on the
+      // next app start.
+      unawaited(_remoteConfig.fetch());
     } catch (e) {
       log("Error fetch remote config: $e");
     }