@nilsreichardt 2023-07-26
> Couldn't initialize be synchronous now? void initialize(...)

Yes, good catch!